Update the ISO builder.
The motivation for all of this is that I want a USB key containing a
GUIX installer, with the Nonguix kernel, that also contains my full
system & user Guix config files, so that I can install all my systems
from a single key, without needing ssh keys or git to clone my config
repo.  Currently the default nonguix ISO doesn't have enough disk space
to build all the things I need in /tmp and it continually runs out of
diskspace.

The upstream repo for this fork hasn't been building ISO's for a while
now. So I forked it in order to fix it, and make some updates.

This commit is a squash of some work I've been doing on this repo
because I made a bad job of the commit history: it's non-atomic and
disordered. I want my changes to be easily distinguishable from the
original branch.

Things that have been changed in this fork:

- The dependency on the guix-install-action builder has been removed in
  favour of installing guix manually on the runner
- Some unused language toolchains have been removed from the runner in
  order to free up as much diskspace as possible for a kernel build in
  /tmp
- The ISO build action has been refactored to rely on the build-iso.sh
  script rather than having similar functionality duplicated across 2
  places.
- The ISO generated has been made larger and will require a larger USB
  key, This is so that the install environment has a larger /tmp space
  for large compiles when using an existing system.scm that compiles
  more software than the basic system.scm provided by the installer does
  (particularly the nonguix kernel).
- An option has been added to skip all substitute servers, and build
  everything from source. Consequently the Github Actions runner timeout
  has been increased to 12 hours (I've not actually tested that this
  timeout is enough for a full build however).
- The action will now check the liveness of the ci and bordeaux
  substitute servers, and will replace them with mirrors if any are not
  responding (NOTE: As of this commit ci and bordeaux are both returning
  504 error codes so the action is hardcoding the fallbacks, this will
  be fixed in a future commit when ci and bordeaux are back online
- An Update script has been added that can be run to update the commits
  of the main Guix repo and the Nonguix repo to their latest commits.
  This will update the main channels.scm file which can then be
  committed back to the repo.